+++ hackathon/xiaoxiaol/consensus_skeleton_2/consensus_skeleton.h
@@ -24,7 +24,7 @@ bool  remove_outliers(vector<NeuronTree> & nt_list, QString SelectedNeuronsAnoFi
 void  isIsolated(vector<int>& isolated,vector<double>& nt_lens,vector<double>& nt_N_bifs, double radius, int count);
 void  normalizeVector(vector<double>& original, vector<double>& normalized);
 void  buildDistanceMatrix(vector< vector<double> >& dist, vector<double>& x, vector<double>& y);
-//void build_tree_from_adj_matrix(QList<NeuronSWC> node_list, double * adjMatrix, QList<NeuronSWC> merge_result);
+//void build_tree_from_adj_matrix(QList<NeuronSWC> node_list, unsigned short * adjMatrix, QList<NeuronSWC> merge_result);
 
 //votemap MST
 double  computeTotalLength(const NeuronTree & nt);
@@ -56,10 +56,10 @@ int cluster_distance_threshold, int resample_flag, V3DPluginCallback2 &callback)
 bool merge_and_vote(vector<NeuronTree> & nt_list_resampled,
                     int vote_threshold,QList<NeuronSWC> &merge_result,int TYPE_MERGED);
 double correspondingNodeFromNeuron(XYZ pt,QList<NeuronSWC> listNodes, int &closestNodeIdx,int TYPE_MERGED);
-bool build_adj_matrix( vector<NeuronTree>  nt_list, QList<NeuronSWC>  merge_result, double * adjMatrix,int TYPE_MERGED);
-int postprocessing_neuron_node_list(QList<NeuronSWC>  &merge_result, double*&  adjMatrix, double vote_threshold);
-bool build_tree_from_adj_matrix(double * adjMatrix, QList<NeuronSWC> &merge_result,double vote_threshold);
-bool build_tree_from_adj_matrix_mst(double * adjMatrix, QList<NeuronSWC> &merge_result, double vote_threshold);
+bool build_adj_matrix( vector<NeuronTree>  nt_list, QList<NeuronSWC>  merge_result, unsigned short * adjMatrix,int TYPE_MERGED);
+int postprocessing_neuron_node_list(QList<NeuronSWC>  &merge_result, unsigned short*&  adjMatrix, double vote_threshold);
+bool build_tree_from_adj_matrix(unsigned short * adjMatrix, QList<NeuronSWC> &merge_result,double vote_threshold);
+bool build_tree_from_adj_matrix_mst(unsigned short * adjMatrix, QList<NeuronSWC> &merge_result, double vote_threshold);
 void trim_unconfident_branches(QList<NeuronSWC> &merge_result,float threshold);
 void generate_batch_trimmed_results(NeuronTree nt,QString outfileName,double initial_threshold, int steps);
 #endif
